% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_ev_model}
\alias{california_ev_model}
\title{EV model example}
\format{
\subsection{\code{california_ev_model}}{

An \code{evmodel} object.
\describe{
\item{metadata}{Information about the characteristics of the model}
\item{model}{Gaussian Mixture Models for connection times and energy}
}
}
}
\source{
\url{https://mcanigueral.github.io/evprof/articles/california.html}
}
\usage{
california_ev_model
}
\description{
Example of an \code{evmodel} object created with \code{{evprof}} for testing purposes.
It has been created using an Open source data set of EV charging sessions
provided by \href{https://acnportal.readthedocs.io/en/latest/}{ACN}.
More information about the development of the model in the {evprof} website:
\url{https://mcanigueral.github.io/evprof/articles/california.html}
}
\keyword{internal}
