% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_ev_sessions_profiles}
\alias{california_ev_sessions_profiles}
\title{Clustered EV charging sessions example}
\format{
\subsection{\code{california_ev_sessions_profiles}}{

A \code{tibble} object with standard variable names defined by \code{evprof}
}
}
\source{
\url{https://ev.caltech.edu/dataset}
}
\usage{
california_ev_sessions_profiles
}
\description{
Example of an charging sessions data set that has been clustered by \code{{evprof}} functions.
(see \href{https://mcanigueral.github.io/evprof/articles/california.html#sessions-classification-into-user-profiles}{this article}).
}
\keyword{internal}
