% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_GMM}
\alias{california_GMM}
\title{Gaussian Mixture Models examples}
\format{
list of two tibbles
\describe{
\item{connection_models}{Tibble with the parameters of the bi-variate
(connection start time and connection duration) GMM from the working days
sessions of the California data set}
\item{energy_models}{Tibble with the parameters of the uni-variate
(energy) GMM from the working days sessions of the California data set}
}
}
\source{
\url{https://mcanigueral.github.io/evprof/articles/california.html}
}
\usage{
california_GMM
}
\description{
Example of connection and energy GMM obtained from functions
\code{get_connection_models} and \code{get_energy_models} respectively.
They have been created using an Open source data set of EV charging sessions
provided by \href{https://acnportal.readthedocs.io/en/latest/}{ACN}.
More information about the development of the model in the evprof website:
\url{https://mcanigueral.github.io/evprof/articles/california.html}
}
\keyword{internal}
