% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_dt_to_num}
\alias{convert_time_dt_to_num}
\title{Convert datetime value to numeric (hour-based)}
\usage{
convert_time_dt_to_num(time_dt)
}
\arguments{
\item{time_dt}{Datetime value}
}
\description{
Convert datetime value to numeric (hour-based)
}
