% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_dt_to_plot_dt}
\alias{convert_time_dt_to_plot_dt}
\title{Modify datetime values according evprof.start.hour}
\usage{
convert_time_dt_to_plot_dt(time_dt, start = getOption("evprof.start.hour"))
}
\arguments{
\item{time_dt}{Datetime value}

\item{start}{Start hour (int)}
}
\description{
Modify datetime values according evprof.start.hour
}
