% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_dt_to_plot_factor}
\alias{convert_time_dt_to_plot_factor}
\title{Convert datetime value to sorted factor considering a start time}
\usage{
convert_time_dt_to_plot_factor(
  time_dt,
  interval = 0.5,
  start = getOption("evprof.start.hour")
)
}
\arguments{
\item{time_dt}{Datetime value}

\item{interval}{Time interval of the time sequence. It can be 0.5 (30 minutes) or 1 (1 hour).}

\item{start}{Start hour (int)}
}
\description{
Convert datetime value to sorted factor considering a start time
}
