% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_dt_to_plot_num}
\alias{convert_time_dt_to_plot_num}
\title{Convert datetime values to sorted numeric values considering a start time}
\usage{
convert_time_dt_to_plot_num(time_dt, start = getOption("evprof.start.hour"))
}
\arguments{
\item{time_dt}{Datetime value}

\item{start}{Start hour (int)}
}
\description{
Convert datetime values to sorted numeric values considering a start time
}
\keyword{internal}
