% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_dt_to_rounded_dt}
\alias{convert_time_dt_to_rounded_dt}
\title{Convert datetime value to rounded numeric (hour-based)}
\usage{
convert_time_dt_to_rounded_dt(time_dt, interval = 0.5)
}
\arguments{
\item{time_dt}{Datetime value}

\item{interval}{Time interval of the time sequence. It can be 0.5 (30 minutes) or 1 (1 hour).}
}
\description{
Convert datetime value to rounded numeric (hour-based)
}
