% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_num_to_chr}
\alias{convert_time_num_to_chr}
\title{Convert numeric time value (hour-based) to character hour in \%H:\%M format}
\usage{
convert_time_num_to_chr(time_num)
}
\arguments{
\item{time_num}{Numeric time value (hour-based)}
}
\description{
Convert numeric time value (hour-based) to character hour in \%H:\%M format
}
\keyword{internal}
