% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_num_to_period}
\alias{convert_time_num_to_period}
\title{Convert numeric time value to a datetime period (hour-based)}
\usage{
convert_time_num_to_period(time_num)
}
\arguments{
\item{time_num}{Numeric time value (hour-based)}
}
\description{
Convert numeric time value to a datetime period (hour-based)
}
\keyword{internal}
