% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{convert_time_num_to_plot_num}
\alias{convert_time_num_to_plot_num}
\title{Modify numeric time value according to a time sequence start time}
\usage{
convert_time_num_to_plot_num(time_num, start = getOption("evprof.start.hour"))
}
\arguments{
\item{time_num}{Numeric time value (hour-based)}

\item{start}{Start hour (int)}
}
\description{
Modify numeric time value according to a time sequence start time
}
