% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{define_profiles}
\alias{define_profiles}
\title{Obtain total sessions data set with Profile classification}
\usage{
define_profiles(sessions_clustered = list(), clusters_interpretations = list())
}
\arguments{
\item{sessions_clustered}{list with sessions clustered of each susbset}

\item{clusters_interpretations}{tibble with clusters interpretations of each subset}
}
\value{
tibble
}
\description{
Obtain total sessions data set with Profile classification
}
