% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a data.frame or tibble to timeseries data.frame}
\usage{
df_to_ts(df)
}
\arguments{
\item{df}{data.frame or tibble}
}
\value{
timeseries data.frame
}
\description{
Convert a data.frame or tibble to timeseries data.frame
}
