% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{divide_by_timecycle}
\alias{divide_by_timecycle}
\title{Divide sessions by time-cycle}
\usage{
divide_by_timecycle(
  sessions,
  wdays_cycles = list(1:5, 6:7),
  months_cycles = list(1:12)
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{wdays_cycles}{list containing Weekdays cycles}

\item{months_cycles}{list containing Monthly cycles}
}
\value{
same sessions data set with extra column "Timecycle"
}
\description{
Divide sessions by time-cycle
}
