% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_charging_rates_distribution}
\alias{get_charging_rates_distribution}
\title{Get charging rates distribution in percentages}
\usage{
get_charging_rates_distribution(sessions, unit = "year")
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{unit}{lubridate \code{floor_date} unit parameter}
}
\value{
tibble
}
\description{
Get charging rates distribution in percentages
}
