% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_demand}
\alias{get_demand}
\title{Obtain timeseries demand from sessions dataset}
\usage{
get_demand(sessions, dttm_seq, by = "Profile")
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{dttm_seq}{vector with a sequence of datetime values}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}
}
\value{
tibble
}
\description{
Obtain timeseries demand from sessions dataset
}
