% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_division_line}
\alias{get_division_line}
\title{{ggplot2} type function to plot a division line}
\usage{
get_division_line(day_n, start_hour)
}
\arguments{
\item{day_n}{Number of the day below the line}

\item{start_hour}{Starting hour}
}
\value{
ggplot2 function
}
\description{
{ggplot2} type function to plot a division line
}
