% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Get ellipse values of a Gaussian model}
\usage{
get_ellipse(mu, sigma)
}
\arguments{
\item{mu}{mean of the g¡Gaussian model}

\item{sigma}{variance of the Gaussian model}
}
\description{
Get ellipse values of a Gaussian model
}
