% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_model}
\alias{get_energy_model}
\title{Get energy univariate Gaussian Mixture Model}
\usage{
get_energy_model(energy_vct, n_models)
}
\arguments{
\item{energy_vct}{energy numeric vector}

\item{n_models}{number of models}
}
\value{
tibble
}
\description{
Get energy univariate Gaussian Mixture Model
}
