% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_model_mclust_object}
\alias{get_energy_model_mclust_object}
\title{Get Mclust object of univariate Gaussian Mixture Models}
\usage{
get_energy_model_mclust_object(energy_vct, log = TRUE)
}
\arguments{
\item{energy_vct}{numeric vector, energy from sessions}

\item{log}{logical, true to perform logarithmic transformation (base = exp(1))}
}
\value{
object of class \code{dnstyMcl}
}
\description{
Get Mclust object of univariate Gaussian Mixture Models
}
