% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_model_parameters}
\alias{get_energy_model_parameters}
\title{Get energy univariate Gaussian Mixture Model}
\usage{
get_energy_model_parameters(mclust_obj)
}
\arguments{
\item{mclust_obj}{object of class \code{dnstyMcl} from function \code{get_energy_model_mclust_object}}
}
\value{
tibble
}
\description{
This function outputs a similar ellipses plot than function \code{plot_bivarGMM()}
but using a different color for each user profile instead of clusters
(the clusters of a same profile have the same color now).
}
\keyword{internal}
