% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_eps}
\alias{get_eps}
\title{Get optimal value of eps according to MinPts kNNdist}
\usage{
get_eps(sessions, MinPts)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{MinPts}{MinPts value}
}
\value{
numeric value
}
\description{
Get optimal value of eps according to MinPts kNNdist
}
