% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_interval_demand}
\alias{get_interval_demand}
\title{Obtain demand from a starting dttm value and certain duration interval}
\usage{
get_interval_demand(
  sessions,
  dttm_start,
  interval_mins,
  by = c("Profile", "Session")
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{dttm_start}{datetime value starting the demand calculation}

\item{interval_mins}{numeric, duration (in minutes) of the demand interval}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}
}
\value{
tibble
}
\description{
Obtain demand from a starting dttm value and certain duration interval
}
