% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_mclust_object}
\alias{get_mclust_object}
\title{Perform \code{mclust::Mclust} clustering}
\usage{
get_mclust_object(
  sessions,
  k,
  mclust_tol = 1e-08,
  mclust_itmax = 10000,
  log = TRUE
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{k}{number of clusters}

\item{mclust_tol}{tolerance parameter for clustering}

\item{mclust_itmax}{maximum number of iterations}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
mclust object
}
\description{
Perform \code{mclust::Mclust} clustering
}
