% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_mclust_params}
\alias{get_mclust_params}
\title{Extract models parameters from mclust object}
\usage{
get_mclust_params(mclust_obj)
}
\arguments{
\item{mclust_obj}{\code{mclust::Mclust} object}
}
\description{
Extract models parameters from mclust object
}
