% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_MinPts}
\alias{get_MinPts}
\title{Get MinPts value according to percentage of sessions data set size (rows)}
\usage{
get_MinPts(sessions, pct = 0.001)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{pct}{percentage of rows to consider}
}
\value{
integer value
}
\description{
Get MinPts value according to percentage of sessions data set size (rows)
}
