% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{get_sessions_profiling}
\alias{get_sessions_profiling}
\title{Obtain total sessions data set with Profile classification}
\usage{
get_sessions_profiling(
  sessions_clustered = list(),
  clusters_interpretations = list()
)
}
\arguments{
\item{sessions_clustered}{list with sessions clustered of each susbset}

\item{clusters_interpretations}{tibble with clusters interpretations of each subset}
}
\value{
tibble
}
\description{
Obtain total sessions data set with Profile classification
}
