% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{mutate_to_log}
\alias{mutate_to_log}
\title{Logarithmic transforamtion to ConnectionStartDateTime and ConnectionHours variables}
\usage{
mutate_to_log(sessions, base = exp(1))
}
\arguments{
\item{sessions}{sessions data set in standard format.}

\item{base}{logarithmic base}
}
\description{
Logarithmic transforamtion to ConnectionStartDateTime and ConnectionHours variables
}
