% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_bivarGMM}
\alias{plot_bivarGMM}
\title{Plot Bivariate Gaussian Mixture Models}
\usage{
plot_bivarGMM(
  sessions,
  bivarGMM_params,
  profiles_names = seq(1, nrow(bivarGMM_params)),
  points_size = 0.5,
  lines_size = 1,
  legend_nrow = 2,
  log = TRUE
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{bivarGMM_params}{parameters of the clusters' GMM models}

\item{profiles_names}{names of profiles}

\item{points_size}{size of scatter points in the plot}

\item{lines_size}{size of lines in the plot}

\item{legend_nrow}{number of rows in legend}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
ggplot2 plot
}
\description{
Plot Bivariate Gaussian Mixture Models
}
