% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_division_lines}
\alias{plot_division_lines}
\title{Iteration over evprof::plot_division_line function to plot multiple lines}
\usage{
plot_division_lines(ggplot_points, n_lines, start_hour)
}
\arguments{
\item{ggplot_points}{ggplot2 returned by evprof::plot_points function}

\item{n_lines}{number of lines to plot}

\item{start_hour}{Starting hour}
}
\value{
ggplot2 function
}
\description{
Iteration over evprof::plot_division_line function to plot multiple lines
}
