% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_divisions}
\alias{plot_divisions}
\title{Plot divisions in a wrapped ggplot}
\usage{
plot_divisions(
  sessions,
  by_disconnection = TRUE,
  by_timecycle = TRUE,
  plottype = c("points", "density"),
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{by_disconnection}{boolean}

\item{by_timecycle}{boolean}

\item{plottype}{"density" or "points" plot}

\item{nrow}{number of rows for the ggplot2::facet_wrap}

\item{ncol}{number of columns for the ggplot2::facet_wrap}
}
\value{
ggplot
}
\description{
Plot divisions in a wrapped ggplot
}
