% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_energy_model}
\alias{plot_energy_model}
\title{Plot energy model}
\usage{
plot_energy_model(profile, mclust_obj)
}
\arguments{
\item{profile}{profile name}

\item{mclust_obj}{object of class \code{dnstyMcl} from function \code{get_energy_model_mclust_object}}
}
\value{
ggplot
}
\description{
Plot energy model
}
