% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_energy_models_density}
\alias{plot_energy_models_density}
\title{Compare density of estimated energy with density of real energy vector}
\usage{
plot_energy_models_density(sessions_profiles, energy_models)
}
\arguments{
\item{sessions_profiles}{sessions data set with user profile attribute}

\item{energy_models}{energy models returned by function \code{get_energy_models}}
}
\value{
list of ggplots
}
\description{
Compare density of estimated energy with density of real energy vector
}
