% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_estimated_energy_density}
\alias{plot_estimated_energy_density}
\title{Plot estimated energy density}
\usage{
plot_estimated_energy_density(profile, energy_vct, estimated_energy)
}
\arguments{
\item{profile}{profile name}

\item{energy_vct}{energy numeric vector}

\item{estimated_energy}{estimated energy numeric vector}
}
\value{
ggplot
}
\description{
Plot estimated energy density
}
