% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_estimated_energy_models}
\alias{plot_estimated_energy_models}
\title{DEPRECATED: Compare Gaussian Mixture Models of estimated energy with density of real energy vector}
\usage{
plot_estimated_energy_models(profile, energy_vct, estimated_energy)
}
\arguments{
\item{profile}{profile name}

\item{energy_vct}{energy numeric vector}

\item{estimated_energy}{estimated energy numeric vector}
}
\value{
ggplot
}
\description{
DEPRECATED: Compare Gaussian Mixture Models of estimated energy with density of real energy vector
}
