% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram of a variable from sessions data set}
\usage{
plot_histogram(sessions, var, binwidth = 1)
}
\arguments{
\item{sessions}{tibble, sessions data set in evprof
\href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{standard format}.}

\item{var}{character, column name to compute the histogram for}

\item{binwidth}{integer, with of histogram bins}
}
\value{
ggplot plot
}
\description{
Histogram of a variable from sessions data set
}
\examples{
plot_histogram(california_ev_sessions, "Power", binwidth = 2)
plot_histogram(california_ev_sessions, "Power", binwidth = 0.1)

}
