% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_kNNdist}
\alias{plot_kNNdist}
\title{Plot kNNdist}
\usage{
plot_kNNdist(sessions, MinPts = NULL, MinPts_pct = 0.001, log = FALSE)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{MinPts}{MinPts value}

\item{MinPts_pct}{MinPts percentage}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
plot
}
\description{
Plot kNNdist
}
