% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_model_clusters}
\alias{plot_model_clusters}
\title{Aggregate clusters GMM into profiles GMM}
\usage{
plot_model_clusters(
  subsets_clustering = list(),
  clusters_interpretations = list(),
  profiles_ratios,
  log = TRUE
)
}
\arguments{
\item{subsets_clustering}{list with clustering results of each subset to aggregate}

\item{clusters_interpretations}{list with clusters interpretations of each subset}

\item{profiles_ratios}{tibble with columns \code{profile} and \code{profile_ratio}}

\item{log}{Logical, true to perform logarithmic transformation (base = exp(1))}
}
\value{
ggplot2
}
\description{
Aggregate clusters GMM into profiles GMM
}
