% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outlying sessions}
\usage{
plot_outliers(sessions)
}
\arguments{
\item{sessions}{sessions data set in standard format}
}
\value{
ggplot2 plot
}
\description{
Plot outlying sessions
}
