% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print_connection_models_table}
\alias{print_connection_models_table}
\title{Get LaTeX code for the connection bivariate GMM features (mu and sigma)}
\usage{
print_connection_models_table(GMM)
}
\arguments{
\item{GMM}{Gaussian Mixture Models obtained from function \code{get_connection_models}}
}
\value{
character, LaTeX code
}
\description{
Get LaTeX code for the connection bivariate GMM features (mu and sigma)
}
