% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print_energy_models_table}
\alias{print_energy_models_table}
\title{Get LaTeX code for the energy GMM features (mu and sigma)}
\usage{
print_energy_models_table(GMM)
}
\arguments{
\item{GMM}{Gaussian Mixture Models obtained from function \code{get_energy_models}}
}
\value{
character, LaTeX code
}
\description{
Get LaTeX code for the energy GMM features (mu and sigma)
}
