% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{read_ev_model}
\alias{read_ev_model}
\title{Read an EV model JSON file and convert it to object of class \code{evmodel}}
\usage{
read_ev_model(file)
}
\arguments{
\item{file}{path to the JSON file}
}
\value{
object of class \code{evmodel}
}
\description{
Read an EV model JSON file and convert it to object of class \code{evmodel}
}
