% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{round_to_half}
\alias{round_to_half}
\title{Round numeric time value to half hour basis.}
\usage{
round_to_half(time_num)
}
\arguments{
\item{time_num}{Numeric time value (hour-based)}
}
\description{
Round numeric time value to half hour basis.
}
