% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{save_clustering_iterations}
\alias{save_clustering_iterations}
\title{Save iteration plots in PDF file}
\usage{
save_clustering_iterations(
  sessions,
  k,
  it = 12,
  seeds = round(runif(it, min = 1, max = 1000)),
  filename = paste0("iteration_", k, "_clusters.pdf"),
  plot_scale = 2,
  points_size = 0.25,
  mclust_tol = 1e-08,
  mclust_itmax = 10000,
  log = TRUE
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{k}{number of clusters}

\item{it}{number of iterations}

\item{seeds}{seed for each iteration}

\item{filename}{PDF output file (with extension .pdf)}

\item{plot_scale}{scale of each iteration plot for a good visualization in pdf file}

\item{points_size}{integer, size of points in the scatter plot}

\item{mclust_tol}{tolerance parameter for clustering}

\item{mclust_itmax}{maximum number of iterations}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
description No returns but a PDF file is saved in the path specified by parameter \code{filename}
}
\description{
Save iteration plots in PDF file
}
