% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_ev_model}
\alias{save_ev_model}
\title{Save the EV model object of class \code{evmodel} to a JSON file}
\usage{
save_ev_model(evmodel, filename = "evmodel.json")
}
\arguments{
\item{evmodel}{object of class \code{evmodel}
(see this \href{https://mcanigueral.github.io/evprof/articles/evmodel.html}{link} for more information)}

\item{filename}{character string with the path or name of the file}
}
\description{
Save the EV model object of class \code{evmodel} to a JSON file
}
