% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_models}
\alias{save_models}
\title{Save the models as .RDS file}
\usage{
save_models(
  connection_GMM = list(),
  energy_GMM = list(),
  models_names,
  months = list(1:12, 1:12),
  wdays = list(1:5, 6:7),
  connection_log = TRUE,
  energy_log = TRUE,
  file = "."
)
}
\arguments{
\item{connection_GMM}{list of different connection bivariate GMM}

\item{energy_GMM}{list of different energy univariate GMM}

\item{models_names}{character vector with the labels of each model}

\item{months}{integer vector with the corresponding months of the year for each model}

\item{wdays}{integer vector with the corresponding days of the week for each model (week start = 1)}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}

\item{file}{character string with the path or name of the .RDS file}
}
\description{
Save the models as .RDS file
}
