% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sessions}
\alias{sessions}
\title{Sample of EV sessions dataset}
\format{
A tibble
}
\usage{
sessions
}
\description{
A dataset containing the features of EV sessions in public
charging infrastructura from September to December 2019,
both included.
}
\keyword{datasets}
