% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sessions_feature_names}
\alias{sessions_feature_names}
\title{Names of standard features of a sessions dataset}
\format{
A vector
}
\usage{
sessions_feature_names
}
\description{
A vector with the standard names of sessions features
for this package functions.
}
\keyword{internal}
