% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sessions_summary_feature_names}
\alias{sessions_summary_feature_names}
\title{Names of features to summarise in {evprof} functions}
\format{
A vector
}
\usage{
sessions_summary_feature_names
}
\description{
A vector with the summary features
}
\keyword{internal}
