% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{set_profiles}
\alias{set_profiles}
\title{Classify sessions of sub-sets into profiles}
\usage{
set_profiles(sessions_clustered = list(), clusters_definition = list())
}
\arguments{
\item{sessions_clustered}{list of tibbles with sessions clustered (output from function \code{cluser_sessions}) from each sub-set}

\item{clusters_definition}{list of tibbles with clusters definitions (output from function \code{define_clusters}) of each sub-set}
}
\value{
tibble
}
\description{
Joins all sub-sets from the list, adding a new column \code{Profile}
}
