% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{summarise_sessions}
\alias{summarise_sessions}
\title{Statistic summary of sessions features}
\usage{
summarise_sessions(
  sessions,
  .funs,
  vars = evprof::sessions_summary_feature_names
)
}
\arguments{
\item{sessions}{tibble, sessions data set in evprof
\href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{standard format}.
\href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{standard format}.}

\item{.funs}{A function to compute, e.g. \code{mean}, \code{max}, etc.}

\item{vars}{character vector, variables to compute the histogram for}
}
\value{
Summary table
}
\description{
Statistic summary of sessions features
}
\examples{
summarise_sessions(california_ev_sessions, mean)


}
