% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{tabulate_clusters_interpretation}
\alias{tabulate_clusters_interpretation}
\title{Tabulate clusters interpretation}
\usage{
tabulate_clusters_interpretation(bivarGMM, interpretations, profile_names)
}
\arguments{
\item{bivarGMM}{clusters models}

\item{interpretations}{vector with interpretation sentences of each cluster (arranged by cluster number)}

\item{profile_names}{vector with user profile assigned to each cluster (arranged by cluster number)}
}
\value{
tibble object
}
\description{
Tabulate clusters interpretation
}
