% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{choose_k_GMM}
\alias{choose_k_GMM}
\title{Visuallize BIC indicator to choose the number of clusters}
\usage{
choose_k_GMM(sessions, k, mclust_tol = 1e-08, mclust_itmax = 10000)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{k}{sequence with the number of clusters, for example 1:10, for 1 to 10 clusters.}

\item{mclust_tol}{tolerance parameter for clustering}

\item{mclust_itmax}{maximum number of iterations}
}
\value{
BIC plot
}
\description{
Visuallize BIC indicator to choose the number of clusters
}
