% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_sessions}
\alias{cluster_sessions}
\title{Cluster sessions with \code{mclust} package}
\usage{
cluster_sessions(
  sessions,
  k,
  seed,
  mclust_tol = 1e-08,
  mclust_itmax = 10000,
  log = TRUE
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{k}{number of clusters}

\item{seed}{random seed}

\item{mclust_tol}{tolerance parameter for clustering}

\item{mclust_itmax}{maximum number of iterations}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
list with two attributes: sessions and models
}
\description{
Cluster sessions with \code{mclust} package
}
