% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{define_clusters}
\alias{define_clusters}
\title{Tabulate clusters interpretation}
\usage{
define_clusters(
  models,
  interpretations = NULL,
  profile_names = NULL,
  log = TRUE
)
}
\arguments{
\item{models}{clusters bivariate Gaussian mixture models}

\item{interpretations}{vector with interpretation sentences of each cluster (arranged by cluster number)}

\item{profile_names}{vector with user profile assigned to each cluster (arranged by cluster number)}

\item{log}{Logical. Whether ConnectionStartDateTime and ConnectionHours variables are in natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
tibble object
}
\description{
Tabulate clusters interpretation
}
