% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(sessions, MinPts, eps)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{MinPts}{MinPts parameter for DBSCAN clustering}

\item{eps}{eps parameter for DBSCAN clustering}
}
\value{
sessions data set with extra boolean column \code{Outlier}
}
\description{
Detect outliers
}
