% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{divide_by_disconnection}
\alias{divide_by_disconnection}
\title{Divide sessions by disconnection day}
\usage{
divide_by_disconnection(sessions, days, start_hour)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{days}{number of disconnection days to select}

\item{start_hour}{Starting hour}
}
\value{
same sessions data set with extra column "Disconnection"
}
\description{
Divide sessions by disconnection day
}
