% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{drop_outliers}
\alias{drop_outliers}
\title{Drop outliers}
\usage{
drop_outliers(sessions)
}
\arguments{
\item{sessions}{sessions data set in standard format}
}
\value{
sessions without outliers nor column \code{Outlier}
}
\description{
Drop outliers
}
