% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_charging_rates_distribution}
\alias{get_charging_rates_distribution}
\title{Get charging rates distribution in percentages}
\usage{
get_charging_rates_distribution(sessions, unit = "year")
}
\arguments{
\item{sessions}{tibble, sessions data set in evprof
\href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{standard format}.}

\item{unit}{character, lubridate \code{floor_date} unit parameter}
}
\value{
tibble
}
\description{
Get charging rates distribution in percentages
}
\examples{
get_charging_rates_distribution(california_ev_sessions, unit="month")
get_charging_rates_distribution(california_ev_sessions, unit="month")

}
