% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_connection_models}
\alias{get_connection_models}
\title{Get a tibble of connection GMM for every user profile}
\usage{
get_connection_models(
  subsets_clustering = list(),
  clusters_interpretations = list()
)
}
\arguments{
\item{subsets_clustering}{list with clustering results of each subset to aggregate}

\item{clusters_interpretations}{list with clusters interpretations of each subset}
}
\value{
tibble
}
\description{
Get a tibble of connection GMM for every user profile
}
