% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_daily_avg_n_sessions}
\alias{get_daily_avg_n_sessions}
\title{Get the daily average number of sessions given a range of years, months and weekdays}
\usage{
get_daily_avg_n_sessions(sessions, years, months, wdays)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{years}{range of years to consider (int)}

\item{months}{range of months to consider (int)}

\item{wdays}{range of weekdays to consider (int)}
}
\value{
tibble with the number of sessions of each date in the given time period
}
\description{
Get the daily average number of sessions given a range of years, months and weekdays
}
