% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_model}
\alias{get_energy_model}
\title{Get energy univariate Gaussian Mixture Model}
\usage{
get_energy_model(energy_vct, k, maxit = 5000, log = TRUE)
}
\arguments{
\item{energy_vct}{energy numeric vector}

\item{k}{number of univariate Gaussian Mixture Models (int)}

\item{maxit}{maximum number of iterations (int)}

\item{log}{Logical, true to perform logarithmic transformation (base = exp(1))}
}
\value{
tibble
}
\description{
Get energy univariate Gaussian Mixture Model
}
