% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_model_parameters}
\alias{get_energy_model_parameters}
\title{Get energy univariate Gaussian Mixture Model}
\usage{
get_energy_model_parameters(mclust_obj)
}
\arguments{
\item{mclust_obj}{object of class \code{dnstyMcl} from function \code{get_energy_model_mclust_object}}
}
\value{
tibble
}
\description{
Get energy univariate Gaussian Mixture Model
}
