% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_models}
\alias{get_energy_models}
\title{Get a tibble of energy GMM for every user profile}
\usage{
get_energy_models(sessions_profiles, log = TRUE)
}
\arguments{
\item{sessions_profiles}{sessions data set with user profile attribute}

\item{log}{Logical, true to perform logarithmic transformation (base = exp(1))}
}
\value{
tibble
}
\description{
Get a tibble of energy GMM for every user profile
}
