% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_energy_models}
\alias{get_energy_models}
\title{Title}
\usage{
get_energy_models(sessions_profiles, k)
}
\arguments{
\item{sessions_profiles}{sessions data set with user profile attribute}

\item{k}{number of univariate Gaussian Mixture Models}
}
\value{
tibble
}
\description{
Title
}
