% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_ev_model}
\alias{get_ev_model}
\title{Get the EV model object of class \code{evmodel}}
\usage{
get_ev_model(
  names,
  months_lst = list(1:12, 1:12),
  wdays_lst = list(1:5, 6:7),
  connection_GMM,
  energy_GMM,
  connection_log,
  energy_log
)
}
\arguments{
\item{names}{character vector with the given names of each time-cycle model}

\item{months_lst}{list of integer vectors with the corresponding months of the year for each time-cycle model}

\item{wdays_lst}{list of integer vectors with the corresponding days of the week for each model (week start = 1)}

\item{connection_GMM}{list of different connection bivariate GMM}

\item{energy_GMM}{list of different energy univariate GMM}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}
}
\value{
object of class \code{evmodel}
}
\description{
Get the EV model object of class \code{evmodel}
}
