% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{mutate_to_log}
\alias{mutate_to_log}
\title{Logarithmic transformation to ConnectionStartDateTime and ConnectionHours variables}
\usage{
mutate_to_log(sessions, start = getOption("evprof.start.hour"), base = exp(1))
}
\arguments{
\item{sessions}{sessions data set in standard format.}

\item{start}{integer, start hour in the x axis of the plot.}

\item{base}{logarithmic base}
}
\description{
Logarithmic transformation to ConnectionStartDateTime and ConnectionHours variables
}
