% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_density_3D}
\alias{plot_density_3D}
\title{Density plot in 3D, considering Start time and Connection duration as variables}
\usage{
plot_density_3D(
  sessions,
  start = getOption("evprof.start.hour"),
  eye = list(x = -1.5, y = -1.5, z = 1.5),
  log = TRUE
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{start}{start hour (int)}

\item{eye}{list containing x, y and z points of view. Example: \code{list(x = -1.5, y = -1.5, z = 1.5)}#'}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}
}
\value{
plotly plot (html)
}
\description{
Density plot in 3D, considering Start time and Connection duration as variables
}
