% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_division_lines}
\alias{plot_division_lines}
\title{Iteration over evprof::plot_division_line function to plot multiple lines}
\usage{
plot_division_lines(ggplot_points, n_lines, division_hour)
}
\arguments{
\item{ggplot_points}{ggplot2 returned by evprof::plot_points function}

\item{n_lines}{number of lines to plot}

\item{division_hour}{Hour to divide the groups according to disconnection time}
}
\value{
ggplot2 function
}
\description{
Iteration over evprof::plot_division_line function to plot multiple lines
}
\examples{
library(dplyr)
california_ev_sessions \%>\%
  sample_frac(0.05) \%>\%
  plot_points(start = 3) \%>\%
  plot_division_lines(n_lines = 1, division_hour = 5)

}
