% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_energy_models}
\alias{plot_energy_models}
\title{Compare density of estimated energy with density of real energy vector}
\usage{
plot_energy_models(energy_models)
}
\arguments{
\item{energy_models}{energy models returned by function \code{get_energy_models}}
}
\value{
ggplot
}
\description{
Compare density of estimated energy with density of real energy vector
}
