% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram of a variable from sessions data set}
\usage{
plot_histogram(sessions, var, binwidth = 1)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{var}{variable to compute the histogram for}

\item{binwidth}{with of histogram bins}
}
\value{
ggplot plot
}
\description{
Histogram of a variable from sessions data set
}
\examples{
library(dplyr)
sessions <- tibble(Power = rnorm(100, mean = 30))
plot_histogram(sessions, "Power", binwidth = 2)
plot_histogram(sessions, "Power", binwidth = 0.1)

}
