% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_histogram_grid}
\alias{plot_histogram_grid}
\title{Title}
\usage{
plot_histogram_grid(
  sessions,
  vars = evprof::sessions_summary_feature_names,
  binwidths = rep(1, length(vars)),
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{vars}{variables to plot}

\item{binwidths}{binwidths of each variable histogram}

\item{nrow}{number of rows of the plot grid}

\item{ncol}{number of columns of the plot grid}
}
\value{
grid plot
}
\description{
Title
}
