% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_kNNdist}
\alias{plot_kNNdist}
\title{Plot kNNdist}
\usage{
plot_kNNdist(sessions, MinPts = NULL, MinPts_pct = 0.01)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{MinPts}{MinPts value}

\item{MinPts_pct}{MinPts percentage}
}
\value{
plot
}
\description{
Plot kNNdist
}
