% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outlying sessions}
\usage{
plot_outliers(sessions, log = FALSE, ...)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{log}{Logical. Whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = \code{exp(1)}).}

\item{...}{arguments to pass to function ggplot2::plot_point}
}
\value{
ggplot2 plot
}
\description{
Plot outlying sessions
}
