% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print.evmodel}
\alias{print.evmodel}
\title{\code{print} method for \code{evmodel} object class}
\usage{
\method{print}{evmodel}(x, ...)
}
\arguments{
\item{x}{\code{evmodel} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} method for \code{evmodel} object class
}
\keyword{internal}
