% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print_connection_models_table}
\alias{print_connection_models_table}
\title{Get LaTeX code for the connection bivariate GMM features (mu and sigma)}
\usage{
print_connection_models_table(GMM, label, caption, full_width, path)
}
\arguments{
\item{GMM}{Gaussian Mixture Models obtained from function \code{get_connection_models}}

\item{label}{character, e.g. "tab:gmm"}

\item{caption}{character, table caption}

\item{full_width}{logical, if true the "*" will be added next to the "table" tag}

\item{path}{character, file path to write the latex table to. Is must have ".tex" extension.}
}
\value{
character, LaTeX code
}
\description{
Get LaTeX code for the connection bivariate GMM features (mu and sigma)
}
