% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{print_user_profile_energy_models_table}
\alias{print_user_profile_energy_models_table}
\title{Get LaTeX code for the energy GMM features OF A SINGLE USER PROFILE (mu and sigma)}
\usage{
print_user_profile_energy_models_table(
  user_profile_GMM,
  label,
  caption,
  full_width,
  path = NULL
)
}
\arguments{
\item{user_profile_GMM}{Gaussian Mixture Models obtained from function \code{get_energy_models}}

\item{label}{character, e.g. "tab:gmm"}

\item{caption}{character, table caption}

\item{full_width}{logical, if true the "*" will be added next to the "table" tag}

\item{path}{character, file path to write the latex table to. Is must have ".tex" extension.
If it is NULL, then the character string is returned instead of writing a file.}
}
\value{
character, LaTeX code
}
\description{
Get LaTeX code for the energy GMM features OF A SINGLE USER PROFILE (mu and sigma)
}
\keyword{internal}
