% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_ev_model}
\alias{save_ev_model}
\title{Save the EV model object of class \code{evmodel} to an external file}
\usage{
save_ev_model(evmodel, file = "evmodel", fileext = ".RDS")
}
\arguments{
\item{evmodel}{object of class \code{evmodel}}

\item{file}{character string with the path or name of the file}

\item{fileext}{character string with the file extension}
}
\description{
Save the EV model object of class \code{evmodel} to an external file
}
