% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{simulate_sessions}
\alias{simulate_sessions}
\title{Simualte sessions given datetime sequence and models}
\usage{
simulate_sessions(dates, ev_models, charging_rates, interval_mins)
}
\arguments{
\item{dates}{datetime vector with dates to simualte (datetime values with hour set to 00:00)}

\item{ev_models}{profiles models}

\item{charging_rates}{charging rates proportions (tibble)}

\item{interval_mins}{interval of time to round the sessions datetime parameters}
}
\value{
tibble
}
\description{
Simualte sessions given datetime sequence and models
}
