% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_sessions}
\alias{simulate_sessions}
\title{Simualte sessions given datetime sequence and models}
\usage{
simulate_sessions(
  ev_models,
  charging_powers,
  dates,
  interval_mins,
  connection_log = TRUE,
  energy_log = TRUE
)
}
\arguments{
\item{ev_models}{tibble with columns: \code{model_name}, \code{months}, \code{wdays}, \code{models}, \code{n_sessions}
The column \code{models} must be a list of tibbles, while each tibble must have the columns
\code{profile}, \code{profile_ratio} (between 0 and 1), \code{connection_models} and \code{energy_models}}

\item{charging_powers}{charging powers proportions (tibble) with two columns: \code{power} and \code{ratio}.
The powers must be in kW and the ratios between 0 and 1.}

\item{dates}{datetime vector with dates to simualte (datetime values with hour set to 00:00)}

\item{interval_mins}{interval of minutes (integer) to round the sessions datetime variables}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}
}
\value{
tibble
}
\description{
Simualte sessions given datetime sequence and models
}
