% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{summarise_sessions}
\alias{summarise_sessions}
\title{Statistic summary of sessions features}
\usage{
summarise_sessions(
  sessions,
  .funs,
  vars = evprof::sessions_summary_feature_names
)
}
\arguments{
\item{sessions}{sessions data set in standard format}

\item{.funs}{A function \code{fun}}

\item{vars}{variables to compute the histogram for}
}
\value{
Summary table
}
\description{
Statistic summary of sessions features
}
